+++ Sources/Classes/core/JavascriptBridge/MiniAppScriptMessageHandler.swift
@@ -835,7 +835,7 @@ extension MiniAppScriptMessageHandler {
 extension MiniAppScriptMessageHandler {
     func didReceiveMAAnalytics(requestParam: RequestParameters?, callbackId: String){
         guard let requestParamValue = requestParam?.analyticsInfo else {
-            executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: prepareMAJavascriptError(MiniAppJavaScriptError.valueIsEmpty))
+            executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: prepareMAJavascriptError(MiniAppJavaScriptError.unexpectedMessageFormat))
             return
         }
         self.hostAppMessageDelegate?.didReceiveMAAnalytics(analyticsInfo: requestParamValue, completionHandler: { result in